Add database setup
Not making verify script fail if formatting is not successful because
it fails on some generated files, even though they're in the filter
in build gradle files. ktlintcheck task correctly ignores them
but not the format task. So if there are formatting issues that ktlint
is not able to fix on its own, the pipeline would fail and would need
to manually fix those changes.

+++ shared/src/commonMain/kotlin/com/ramitsuri/podcasts/network/api/PodcastsApiImpl.kt
@@ -42,7 +42,7 @@ internal class PodcastsApiImpl(
         }
     }
 
-    override suspend fun getById(id: Int): PodcastResult<PodcastResponseDto> {
+    override suspend fun getById(id: Long): PodcastResult<PodcastResponseDto> {
         return apiRequest(ioDispatcher) {
             httpClient.get(
                 urlString =
