+++ shared/src/commonMain/kotlin/com/ramitsuri/podcasts/network/model/GetEpisodesRequest.kt
@@ -2,4 +2,4 @@ package com.ramitsuri.podcasts.network.model
 
 import kotlinx.datetime.Instant
 
-internal data class GetEpisodesRequest(val id: Int, val sinceTime: Instant)
+internal data class GetEpisodesRequest(val id: Long, val sinceTime: Instant)
